Update resify a bit
Note: some of these changes have literally been sitting in stage for..
uh.. maybe two years? They appear to work?

This is mostly because I just added the markdown filter for HTML
templates, so that way linkify as a filter can basically be considered
dead.

This also includes a go.mod for the sake of tracking what this is
actually using.